authentication no longer creates a new requestor for self
Requestor created when the octane object is created instead of when
first authenticating. This prevents the requestor is not a function
exception from occurring when an Unauthorized response is expected.
Credentials can also be given as an optional parameter.
These two changes mean that when the Unauthorized response is received
from the server, because the credentials were given beforehand, the
request could still succeed even if no explicit authentication is
called by the user.